requirements.txt
Django>=1.7
markdown>=2.5

django-user-accounts>=1.0
django-crispy-forms>=1.4.0

django-environ>=0.3.0

django-admin-bootstrapped>=1.6.9
django-debug-toolbar>=1.2.1
